+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Plutus/Context.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
@@ -17,6 +18,7 @@
 module Cardano.Ledger.Alonzo.Plutus.Context (
   LedgerTxInfo (..),
   EraPlutusTxInfo (..),
+  PlutusTxInfoResult (..),
   EraPlutusContext (..),
   toPlutusWithContext,
   lookupTxInfoResultImpossible,
@@ -38,10 +40,12 @@ import Cardano.Ledger.Alonzo.Era (AlonzoEra)
 import Cardano.Ledger.Alonzo.Scripts (
   AlonzoEraScript (eraMaxLanguage, mkPlutusScript),
   AsIxItem (..),
+  AsPurpose,
   PlutusPurpose,
   PlutusScript (..),
   hoistPlutusPurpose,
   toAsItem,
+  toAsPurpose,
  )
 import Cardano.Ledger.Alonzo.UTxO (AlonzoEraUTxO (getSpendingDatum))
 import Cardano.Ledger.BaseTypes (ProtVer (..))
@@ -81,15 +85,25 @@ import qualified PlutusLedgerApi.V2 as PV2
 import qualified PlutusLedgerApi.V3 as PV3
 
 -- | All information that is necessary from the ledger to construct Plutus' TxInfo.
-data LedgerTxInfo era = LedgerTxInfo
-  { ltiProtVer :: !ProtVer
-  , ltiEpochInfo :: !(EpochInfo (Either Text))
-  , ltiSystemStart :: !SystemStart
-  , ltiUTxO :: !(UTxO era)
-  , ltiTx :: !(Tx TopTx era)
-  }
+data LedgerTxInfo era where
+  LedgerTxInfo ::
+    { ltiProtVer :: !ProtVer
+    , ltiEpochInfo :: !(EpochInfo (Either Text))
+    , ltiSystemStart :: !SystemStart
+    , ltiUTxO :: !(UTxO era)
+    , ltiTx :: !(Tx level era)
+    } ->
+    LedgerTxInfo era
 
-class (PlutusLanguage l, EraPlutusContext era) => EraPlutusTxInfo (l :: Language) era where
+class
+  ( PlutusLanguage l
+  , EraPlutusContext era
+  , EraTxLevel era
+  , Eq (PlutusTxInfo l)
+  , Show (PlutusTxInfo l)
+  ) =>
+  EraPlutusTxInfo (l :: Language) era
+  where
   toPlutusTxCert ::
     proxy l ->
     ProtVer ->
@@ -105,7 +119,7 @@ class (PlutusLanguage l, EraPlutusContext era) => EraPlutusTxInfo (l :: Language
   toPlutusTxInfo ::
     proxy l ->
     LedgerTxInfo era ->
-    Either (ContextError era) (PlutusTxInfo l)
+    PlutusTxInfoResult l era
 
   toPlutusArgs ::
     proxy l ->
@@ -121,6 +135,21 @@ class (PlutusLanguage l, EraPlutusContext era) => EraPlutusTxInfo (l :: Language
     UTxO era ->
     TxIn ->
     Either (ContextError era) (PlutusTxInInfo era l)
+
+-- | This is the helper type that captures translation of `Tx` to `PlutusTxInfo`.
+--
+-- It is important to note that `TxInfo` is always the same per Plutus version for each `Tx`. This
+-- invariant allows us to avoid duplicate computation by memoizing all possible `PlutusTxInfo`s per
+-- transaction. Starting with Dijkstra era there is a slight complication introduced to this
+-- invariant where top level transaction has a different `PlutusTxInfo` for "Guarding" purpose, when
+-- compared to all other purposes. That is the reason why result is somewhat strange, namely a
+-- function from `PlutusPurpose` to `PlutusTxInfo`. It is also done this way, instead of adding
+-- `ScriptPurpose` as an argument to `toPlutusTxInfo` to preserve capability of memoization.
+newtype PlutusTxInfoResult l era
+  = PlutusTxInfoResult
+  { unPlutusTxInfoResult ::
+      Either (ContextError era) (PlutusPurpose AsPurpose era -> PlutusTxInfo l)
+  }
 
 class
   ( AlonzoEraScript era
@@ -154,7 +183,7 @@ class
     EraPlutusTxInfo l era =>
     SLanguage l ->
     TxInfoResult era ->
-    Either (ContextError era) (PlutusTxInfo l)
+    PlutusTxInfoResult l era
 
   mkPlutusWithContext ::
     PlutusScript era ->
@@ -177,11 +206,12 @@ toPlutusWithContext ::
   (Data era, ExUnits) ->
   CostModel ->
   Either (ContextError era) PlutusWithContext
-toPlutusWithContext script scriptHash plutusPurpose lti txInfoResult (redeemerData, exUnits) costModel = do
+toPlutusWithContext script scriptHash plutusPurpose lti@LedgerTxInfo {ltiTx} txInfoResult (redeemerData, exUnits) costModel = do
   let slang = isLanguage @l
       maybeSpendingDatum =
-        getSpendingDatum (ltiUTxO lti) (ltiTx lti) (hoistPlutusPurpose toAsItem plutusPurpose)
-  txInfo <- lookupTxInfoResult slang txInfoResult
+        getSpendingDatum (ltiUTxO lti) ltiTx (hoistPlutusPurpose toAsItem plutusPurpose)
+  mkTxInfo <- unPlutusTxInfoResult $ lookupTxInfoResult slang txInfoResult
+  let txInfo = mkTxInfo $ hoistPlutusPurpose toAsPurpose plutusPurpose
   plutusArgs <-
     toPlutusArgs slang (ltiProtVer lti) txInfo plutusPurpose maybeSpendingDatum redeemerData
   pure $
@@ -197,7 +227,7 @@ toPlutusWithContext script scriptHash plutusPurpose lti txInfoResult (redeemerDa
 -- | Helper function to use when implementing `lookupTxInfoResult` for plutus languages that are not
 -- supported by the era.
 lookupTxInfoResultImpossible ::
-  (HasCallStack, EraPlutusTxInfo l era) => SLanguage l -> Either (ContextError era) (PlutusTxInfo l)
+  (HasCallStack, EraPlutusTxInfo l era) => SLanguage l -> PlutusTxInfoResult l era
 lookupTxInfoResultImpossible slang =
   error $ "Impossible: Attempt to lookup TxInfoResult for an unsupported language: " <> show slang
 
@@ -223,7 +253,7 @@ type family PlutusTxInfo (l :: Language) where
   PlutusTxInfo 'PlutusV4 = PV3.TxInfo
 
 type family PlutusTxInInfo era (l :: Language) where
-  -- \| This special case is here because Alonzo does not have a ContextError
+  -- This special case is here because Alonzo does not have a ContextError
   -- for the case where it encounters a Byron address in a TxIn
   PlutusTxInInfo AlonzoEra PlutusV1 = Maybe PV1.TxInInfo
   PlutusTxInInfo _ 'PlutusV1 = PV1.TxInInfo