+++ eras/conway/impl/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.22.0.0
 
+* Add `ConwayEraScript` superclass to `ConwayEraPlutusTxInfo`
+* Change `transPlutusPurposeV1V2` to work only with `ScriptPurpose` for any era and change it from ``AsItem` to `AsIxItem`.
+* Change `transPlutusPurposeV3` to work only with `PlutusV3`, but with `ScriptPurpose` for any era.
+* `ConwayTxInfoResult` changed its content type to `PlutusTxInfoResult`
 * Remove `ToCBOR` and `FromCBOR` instances for `DefaultVote`
 
 ## 1.21.0.0